Comment. Dependencies cleanup.
