+++ utils/process/Cargo.toml
@@ -15,10 +15,7 @@ lock = ["fs2"]
 
 [dependencies]
 actix = { version = "0.13", default-features = false }
-# actix-rt = "2.7"
 anyhow = "1.0"
-# bytes = { version = "1", optional = true }
-# chrono = "0.4"
 derive_more = { workspace = true }
 futures = "0.3"
 futures-util = "0.3.4"
@@ -28,9 +25,6 @@ log = "0.4"
 shared_child = "0.3.4"
 tokio = { version = "1", features = ["process", "signal"] }
 thiserror = "1.0"
-# serde = { version = "^1.0", features = ["derive"] }
-# serde_json = "1.0"
-# serde_yaml = "0.9"
 
 fs2 = { version = "0.4.3", optional = true }
