+++ utils/process/src/unix/usage.rs
@@ -3,42 +3,6 @@ use std::time::Duration;
 
 use crate::SystemError;
 
-use super::Process;
-
-#[cfg(target_os = "linux")]
-use nix::unistd::sysconf;
-#[cfg(target_os = "linux")]
-use nix::unistd::SysconfVar::CLK_TCK;
-
-impl Process {
-    pub fn usage(pid: i32) -> Result<Usage, SystemError> {
-        let stat = StatStub::read(pid)?;
-        let tps = Self::ticks_per_second()?;
-
-        let cpu_sec = Duration::from_secs((stat.stime + stat.utime) / tps as u64);
-        let rss_gib = stat.rss as f64 / (1024. * 1024.);
-
-        Ok(Usage { cpu_sec, rss_gib })
-    }
-
-    pub fn info(pid: i32) -> Result<Process, SystemError> {
-        let stat = StatStub::read(pid)?;
-        Ok(Process {
-            pid: stat.pid,
-            ppid: stat.ppid,
-            pgid: stat.pgid,
-        })
-    }
-
-    fn ticks_per_second() -> Result<i64, SystemError> {
-        match sysconf(CLK_TCK) {
-            Ok(Some(tps)) => Ok(tps),
-            Ok(None) => Err(nix::errno::Errno::ENOTSUP.into()),
-            Err(err) => Err(err.into()),
-        }
-    }
-}
-
 pub struct Usage {
     pub cpu_sec: Duration,
     pub rss_gib: f64,
@@ -77,7 +41,7 @@ pub fn getrusage(resource: i32) -> Result<Usage, SystemError> {
 
 #[cfg(target_os = "linux")]
 #[derive(Clone, Debug, Eq, PartialEq, Default)]
-struct StatStub {
+pub(super) struct StatStub {
     pub pid: i32,
     pub comm: String,
     pub state: char,
